message_bot/requirements.txt
browser-use
langchain
supabase
python-dotenv
pydantic-ai
google-genai
